requirements.txt
PyYAML~=6.0
pydantic~=1.9.0